Refactoring OrderService: semplificata la struttura e aggiunte due rotte per l'inserimento e la cancellazione dei prodotti legati all'ordine

+++ DummyDb.cs
@@ -1,5 +1,8 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using AddressService.Models.Data;
+using Microsoft.EntityFrameworkCore;
 using OrderService.Models.Data;
+using ProductService.Models.Data;
+using UserService.Models.Data;
 
 namespace OrderManagerApi
 {
@@ -9,87 +12,91 @@ namespace OrderManagerApi
         {
             var ctx = new DummyDbContext();
 
-            ctx.Database.EnsureDeleted();
-            ctx.Database.EnsureCreated();
-
-            using (var tr = ctx.Database.BeginTransaction())
+            var mustCreateDb = false;
+            if (mustCreateDb)
             {
-                var users = new List<User>()
+                ctx.Database.EnsureDeleted();
+                ctx.Database.EnsureCreated();
+
+                using (var tr = ctx.Database.BeginTransaction())
                 {
-                    new User() { Id = 1, Name = "Max", Email = "shaolinlohan@gmail.com", PhoneNumber = "3394408189" }
+                    var users = new List<User>()
+                {
+                    new User() { Id = 1, Name = "Max", Email = "test@gmail.com", PhoneNumber = "3391234567" }
                 };
 
-                ctx.Users.AddRange(users);
+                    ctx.Users.AddRange(users);
 
-                ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Users ON");
-                ctx.SaveChanges();
-                ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Users OFF");
+                    ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Users ON");
+                    ctx.SaveChanges();
+                    ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Users OFF");
 
 
-                var addresses = new List<Address>()
+                    var addresses = new List<Address>()
                 {
                     new Address { Id = 1, UserId = 1, Street = "Via degli orti", StreetNumber = "11", City = "Mandello del Lario", ZipCode = "23826" },
                     new Address { Id = 2, UserId = 1, Street = "Via delle erbe", StreetNumber = "13", City = "Mandello del Lario", ZipCode = "23826" }
                 };
 
-                ctx.Addresses.AddRange(addresses);
+                    ctx.Addresses.AddRange(addresses);
 
-                ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Addresses ON");
-                ctx.SaveChanges();
-                ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Addresses OFF");
+                    ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Addresses ON");
+                    ctx.SaveChanges();
+                    ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Addresses OFF");
 
 
-                var categories = new List<Category>()
+                    var categories = new List<Category>()
                 {
                     new Category() { Id = 1, Name = "Attrezzi" },
                     new Category() { Id = 2, Name = "Abbigliamento" },
                 };
 
-                ctx.Categories.AddRange(categories);
+                    ctx.Categories.AddRange(categories);
 
-                ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Categories ON");
-                ctx.SaveChanges();
-                ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Categories OFF");
+                    ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Categories ON");
+                    ctx.SaveChanges();
+                    ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Categories OFF");
 
-                var products = new List<Product>()
+                    var products = new List<Product>()
                 {
                     new Product() { Id = 1, CategoryId = 1, Name = "Racchetta RF 01", Price = 280 },
                     new Product() { Id = 2, CategoryId = 2, Name = "T-Shirt RF", Price = 50 },
                     new Product() { Id = 3, CategoryId = 2, Name = "Borsone RF Collection", Price = 240 },
                 };
 
-                ctx.Products.AddRange(products);
+                    ctx.Products.AddRange(products);
 
-                ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Products ON");
-                ctx.SaveChanges();
-                ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Products OFF");
+                    ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Products ON");
+                    ctx.SaveChanges();
+                    ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Products OFF");
 
-                var orders = new List<Order>()
+                    var orders = new List<Order>()
                 {
-                    new Order() { Id = 1, UserId = 1, AddressId = 1, OrderDate = DateTime.Now },
-                    new Order() { Id = 2, UserId = 1, AddressId = 1, OrderDate = DateTime.Now }
+                    new Order() { Id = 1, UserId = 1, OrderDate = DateTime.Now },
+                    new Order() { Id = 2, UserId = 2, OrderDate = DateTime.Now }
                 };
 
-                ctx.Orders.AddRange(orders);
+                    ctx.Orders.AddRange(orders);
 
-                ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Orders ON");
-                ctx.SaveChanges();
-                ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Orders OFF");
+                    ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Orders ON");
+                    ctx.SaveChanges();
+                    ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.Orders OFF");
 
-                var ordersProducts = new List<OrderProduct>()
+                    var ordersProducts = new List<OrderProduct>()
                 {
                     new OrderProduct() { Id = 1, OrderId = 1, ProductId = 1, Quantity = 2 },
                     new OrderProduct() { Id = 2, OrderId = 1, ProductId = 2, Quantity = 1 },
                     new OrderProduct() { Id = 3, OrderId = 2, ProductId = 3, Quantity = 1 },
                 };
 
-                ctx.OrderProducts.AddRange(ordersProducts);
+                    ctx.OrderProducts.AddRange(ordersProducts);
 
-                ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.OrdersProducts ON");
-                ctx.SaveChanges();
-                ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.OrdersProducts OFF");
+                    ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.OrdersProducts ON");
+                    ctx.SaveChanges();
+                    ctx.Database.ExecuteSqlRaw("SET IDENTITY_INSERT dbo.OrdersProducts OFF");
 
-                tr.Commit();
+                    tr.Commit();
+                }
             }
         }
